Trying to add _process.
Issue with types.

Node has _process, Resource does not.
Base design of audio in godot has:
	AudioStreamPlayer (Node)
		has-a stream : AudioStream (Resource)
			has-a playback : AudioStreamPlayback (Object)

So I have SoundFont (Resource) which is data from sf2 file.
Then if I want 1 class to do playback, it must be AudioStreamPlayer.
So SoundFontPlayer : AudioStreamPlayer. So it is a Node, and has _process.
Then SoundFontGenerator : AudioStream (Resource). That one has SoundFont as member.

Oh, I should put all the methods in the Resource. That will work.